+++ ft-sdk/src/main/java/com/ft/sdk/garble/utils/PackageUtils.java
@@ -1,5 +1,7 @@
 package com.ft.sdk.garble.utils;
 
+import java.lang.reflect.Field;
+
 /**
  * 根据编辑后 Runtime 类映射来判断，应用程序是否对一个库进行依赖
  *
@@ -7,12 +9,12 @@ package com.ft.sdk.garble.utils;
  */
 public class PackageUtils {
     /**
-     *  native SDK 依赖主类 package 路径
+     * native SDK 依赖主类 package 路径
      */
     private static final String PACKAGE_NATIVE_ENGINE_CLASS = "com.ft.sdk.nativelib.NativeEngineInit";
 
     /**
-     *  Okhttp 主类 package 路径
+     * Okhttp 主类 package 路径
      */
     private static final String PACKAGE_OKHTTP3 = "okhttp3.OkHttpClient";
 
@@ -32,6 +34,7 @@ public class PackageUtils {
 
     /**
      * 是否使用 NDK 库
+     *
      * @return
      */
     public static boolean isNativeLibrarySupport() {
@@ -44,7 +47,29 @@ public class PackageUtils {
     }
 
     /**
+     * 获取 native library 库的版本👌
+     *
+     * @return
+     */
+    public static String getNativeLibVersion() {
+
+        try {
+            Class<?> buildConfigClass = Class.forName("com.ft.sdk.nativelib.BuildConfig");
+            Field versionNameField = buildConfigClass.getField("VERSION_NAME");
+            return (String) versionNameField.get(null);
+        } catch (ClassNotFoundException e) {
+            e.printStackTrace();
+        } catch (NoSuchFieldException e) {
+            e.printStackTrace();
+        } catch (IllegalAccessException e) {
+            e.printStackTrace();
+        }
+        return "";
+    }
+
+    /**
      * 是否依赖 okhttp3
+     *
      * @return
      */
     public static boolean isOKHttp3Support() {
@@ -59,6 +84,7 @@ public class PackageUtils {
 
     /**
      * 获取 SophixApplication 的 class
+     *
      * @return
      */
     public static Class<?> getSophixClass() {
@@ -71,6 +97,7 @@ public class PackageUtils {
 
     /**
      * 判断是否是第三方框架数据接入
+     *
      * @return
      */
     public static boolean isThirdPartySupport() {