+++ ft-sdk/CHANGELOG.md
@@ -1,3 +1,7 @@
+# agent 1.4.3-alpha02
+1. 修正 Native Library 版本号获取方式
+
+---
 # agent 1.4.3-alpha01
 1. Native Crash 捕获时机优化，会尝试在崩溃当下捕获
 2. Java Crash 在崩溃入库后，再进行 rethrow